Move iKey from string resources to AndroidManifest.xml
This change allows the instrumentation key to be set from AndroidManifest.xml. Additionally the change refactors the reader of the instrumentation key (in TelemetryChannelConfig.java) to read from the manifest and the unit tests are updated to load a full activity with access to the manifest. The change also adds support for configuring the instrumentation key via gradle manifestPlaceholders. The old test instrumentation key will be deprecated/deleted and a new one will be used. The documentation has also been updated to relfect this change.