Combine sender and queue config
It doesn't make sense to have these separate even though the sender is instance based the Queue is the only consumer of the sender and the queue is a singleton. This change combines the configurations. Prior to this change there was no way to access the sender config as an SDK consumer.