Fix duplicated sessions in lifecycle tracking
the count was always 0 when onResume was called because onPause is always called first. This moves the logic for sending the session start event into the onCreate method and only fires the first time the method is called. The session end event is removed as there is no deterministic way to track this with Android lifecycle flow. Page views are triggered each time onResume is called. Finally, the persistent session storage is removed as we want to track each app start as a new session.