Add basic uncaught exception handler
This will track uncaught exceptions using the standard exception data type. Support for the crash data type still needs to be implemented to enable deobfuscation. Additionally, this currently transmits as the app is shutting down, this should be modified to write the crash to disk and let the app shutdown without attempting network IO.